FTS-1010: Replace 32 file ids with 64

+++ src/bringonline-daemon/context/DeletionContext.h
@@ -85,7 +85,7 @@ public:
     /**
      * Asynchronous update of a single transfer-file within a job
      */
-    void updateState(const std::string &jobId, int fileId, const std::string &state,
+    void updateState(const std::string &jobId, uint64_t fileId, const std::string &state,
         const JobError &error) const;
 
     /**
